Twitter information of Warrior Forums users
analyze_warriors.py is the script to extract Twitter information of
Warrior Forums users and write the Twitter user ids to file.

Improved the user_streams.py file so that each slave machine only
works on the id's it is responsible for. This is a hack in get_users
function.

get_url_redirects.py no longer fetches the entire webpage, instead
just logs the redirects. Even here, each slave machine now works on
ids it is responsible for.

find_fetched_users.sh is the script to look at the user data files and
extract the users for whom information has already been fetched.