Fix the stream fetching code to exit on C-c
Also the update streams code was spending way too many calls just
getting 0 to few tweets for each user and hence was not being
efficient. Now the code only updates a file which has not been updated
in the last 12 hours. If the file had been updated in the last 12
hours, do not update it.